lesson13_btree/rating.txt+lesson14_stream/request.txt

+++ part02/lesson13_btree/btree.c
@@ -8,14 +8,14 @@ extern "C"
 
 	int main(void)
 	{
-		int leaves = 0, i = 1, height = 0;
-		while ((scanf("%d\n", &leaves)) != EOF)
+		unsigned long leaves = 0, i = 1, height = 0;
+		while ((scanf("%lu\n", &leaves)) != EOF)
 		{
 			for (height = 0, i = 1; i < leaves; height++)
 			{
 				i = i << 1;
 			}
-			printf("%d\n", height);
+			printf("%lu\n", height);
 		}
 		return EXIT_SUCCESS;
 	}
